Noticed prototype doesn't match live
On live, we changed assessment centre to "location" because of SRA product that was meant to replace the HAS microservice. We either need to commit to changing it all back to Assessment centre or stick with location (whichever we decide, we need to be consistent and not imply the developers should change it if we haven't made the decision). I've mainly updated V11 but have edited earlier cancellation reason as well since that is what devs were using to build UI)

GitOrigin-RevId: 9d8d195d3b2ba021f3e272aae619a78b8d060e34